Core/Garrison : fix warning

+++ src/server/scripts/Draenor/Garrison/GarrisonNPC.cpp
@@ -102,7 +102,7 @@ class npc_GarrisonFord : public CreatureScript
             else
             {
                 if (p_Player->GetCurrency(GARRISON_CURRENCY_ID, false))
-                    p_Player->ModifyCurrency(GARRISON_CURRENCY_ID, -p_Player->GetCurrency(GARRISON_CURRENCY_ID, false));
+                    p_Player->ModifyCurrency(GARRISON_CURRENCY_ID, -(int32)p_Player->GetCurrency(GARRISON_CURRENCY_ID, false));
 
                 p_Player->DeleteGarrison();
             }
